+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Added
 - Added deprecation warnings to `Connection::runDdl` and `Connection::runDdls` (#98)
+- Added  `ManagesMutations::insertOrUpdateUsingMutation` and `UsesMutations::insertOrUpdateUsingMutation` to do upserts (#109)
 
 Changed
 - `Connection::waitForOperation` and `Connection::isDoneOperation` has been removed. (#99)